pyproject.toml
[project]
name = "forward-forward"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.1",
    "flax>=0.12.2",
    "ipython>=9.10.0",
    "jax[cuda13]>=0.9.0.1",
    "jupyterlab>=4.5.5",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "optax>=0.2.7",
    "pandas>=3.0.1",
    "scikit-learn>=1.6.1",
    "tensorboard>=2.20.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.2",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["forward_forward"]
